requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.6
django-admin-interface==0.17.0
django-colorfield==0.4.3
django-flat-responsive==2.0
django-flat-theme==1.1.4
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.9.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
whitenoise==5.3.0

